[Refactor] Button 컴포넌트 파일 분리 및 전역 색상 적용
