+++ my-blog-app/src/components/Button/Button.tsx
@@ -1,0 +1,36 @@
+import type { ButtonProps } from "./Button.types";
+import { base, variants } from "./Button.styled";
+
+export function Button({
+  children,
+  icon,
+  iconSize,
+  onClick,
+  variant = "primary",
+  disabled = false,
+}: ButtonProps) {
+  const defaultIconSize =
+    iconSize ?? (variant === "tag" || variant === "tagFilled" ? 10.5 : 18);
+
+  return (
+    <button
+      onClick={onClick}
+      disabled={disabled}
+      className={`${base} ${variants[variant]}`}
+    >
+      {icon && (
+        <span
+          style={{
+            width: `${defaultIconSize}px`,
+            height: `${defaultIconSize}px`,
+            display: "inline-flex",
+            alignItems: "center",
+          }}
+        >
+          {icon}
+        </span>
+      )}
+      {children}
+    </button>
+  );
+}
